+++ src/card/cards/combat_constraints.rs
@@ -101,3 +101,4 @@ pub const SPITTING_SLUG: CardDefinitionId = CardDefinitionId(1900);
 pub const ARCADES_SABBOTH: CardDefinitionId = CardDefinitionId(1901);
 pub const ROOTWALLA: CardDefinitionId = CardDefinitionId(1902);
 pub const STAB_WOUND: CardDefinitionId = CardDefinitionId(1903);
+pub const GIANT_SHARK: CardDefinitionId = CardDefinitionId(1904);