+++ src/card/model/presentation.rs
@@ -77,6 +77,7 @@ fn object_predicate_implies(predicate: ObjectPredicateDef, expected: ObjectPredi
         | ObjectPredicateDef::Source
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
+        | ObjectPredicateDef::WasDealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -127,6 +128,7 @@ fn predicate_color(predicate: ObjectPredicateDef) -> Option<ManaColor> {
         | ObjectPredicateDef::Source
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
+        | ObjectPredicateDef::WasDealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -179,6 +181,7 @@ fn predicate_color_count(predicate: ObjectPredicateDef) -> Option<u8> {
         | ObjectPredicateDef::Source
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
+        | ObjectPredicateDef::WasDealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -232,6 +235,7 @@ fn predicate_subtype(predicate: ObjectPredicateDef) -> Option<&'static str> {
         | ObjectPredicateDef::Source
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
+        | ObjectPredicateDef::WasDealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -287,6 +291,7 @@ fn predicate_negated_subtype(predicate: ObjectPredicateDef) -> Option<&'static s
         | ObjectPredicateDef::Source
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
+        | ObjectPredicateDef::WasDealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -348,6 +353,7 @@ fn predicate_power_at_least(predicate: ObjectPredicateDef) -> Option<i16> {
         | ObjectPredicateDef::Source
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
+        | ObjectPredicateDef::WasDealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -395,6 +401,7 @@ fn predicate_mana_value_at_most(predicate: ObjectPredicateDef) -> Option<u8> {
         | ObjectPredicateDef::Source
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
+        | ObjectPredicateDef::WasDealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -447,6 +454,7 @@ fn predicate_controller(predicate: ObjectPredicateDef) -> Option<PlayerRelation>
         | ObjectPredicateDef::Source
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
+        | ObjectPredicateDef::WasDealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
@@ -502,6 +510,7 @@ fn predicate_negates(predicate: ObjectPredicateDef, expected: ObjectPredicateDef
         | ObjectPredicateDef::Source
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
+        | ObjectPredicateDef::WasDealtDamageThisTurn
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource